# core-lending-delinquency-collections-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-delinquency-collections
  description: Loan Delinquency & Collections Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-collections
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/collection-cases/{collectionCaseId}/status-history/{statusHistoryId}:
    get:
      tags:
        - CollectionStatusHistory
      summary: Get a status history record by ID
      operationId: getById_1
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: statusHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionStatusHistoryDTO'
    put:
      tags:
        - CollectionStatusHistory
      summary: Update a status history record
      operationId: update_1
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: statusHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionStatusHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionStatusHistoryDTO'
    delete:
      tags:
        - CollectionStatusHistory
      summary: Delete a status history record
      operationId: delete_1
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: statusHistoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/collection-cases/{collectionCaseId}/status-history:
    get:
      tags:
        - CollectionStatusHistory
      summary: List or search status history records for a collection case
      operationId: findAll_1
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.collectionCaseId
          in: query
          description: Exact filter for collection case id
          required: false
          schema:
            type: string
        - name: filters.oldStatus
          in: query
          description: Exact filter for old status
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - CLOSED
              - LEGAL
              - RESTRUCTURED
              - SETTLED
        - name: filters.newStatus
          in: query
          description: Exact filter for new status
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - CLOSED
              - LEGAL
              - RESTRUCTURED
              - SETTLED
        - name: filters.reasonCode
          in: query
          description: Exact filter for reason code
          required: false
          schema:
            type: string
            enum:
              - PROMISE_BROKEN
              - LEGAL_REFERRAL
              - PAYMENT_RECEIVED
              - NO_CONTACT
        - name: filters.changedAt
          in: query
          description: Exact filter for changed at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[changedAt].from
          in: query
          description: Filter changed at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[changedAt].to
          in: query
          description: Filter changed at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.changedBy
          in: query
          description: Exact filter for changed by
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCollectionStatusHistoryDTO'
    post:
      tags:
        - CollectionStatusHistory
      summary: Create a new status history entry for a case
      operationId: create_1
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionStatusHistoryDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionStatusHistoryDTO'
...
components:
  schemas:
...
    CollectionStatusHistoryDTO:
      required:
        - changedAt
        - changedBy
        - collectionCaseId
        - newStatus
        - reasonCode
      type: object
      properties:
        collectionStatusHistoryId:
          type: string
          format: uuid
          readOnly: true
        collectionCaseId:
          type: string
          format: uuid
        oldStatus:
          type: string
          enum:
            - OPEN
            - CLOSED
            - LEGAL
            - RESTRUCTURED
            - SETTLED
        newStatus:
          type: string
          enum:
            - OPEN
            - CLOSED
            - LEGAL
            - RESTRUCTURED
            - SETTLED
        reasonCode:
          type: string
          enum:
            - PROMISE_BROKEN
            - LEGAL_REFERRAL
            - PAYMENT_RECEIVED
            - NO_CONTACT
        changedAt:
          type: string
          format: date-time
        changedBy:
          maxLength: 255
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseCollectionStatusHistoryDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CollectionStatusHistoryDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...